model/tests/requirements.txt
os
pandas
numpy
scipy_sparse
pickle
mlflow
matplotlib_pyplot
sklearn_feature_extraction_text
sklearn_metrics_pairwise